+++ experimental/packages/opentelemetry-configuration/src/EnvironmentConfigProvider.ts
@@ -73,8 +73,18 @@ export function setResources(config: ConfigurationModel): void {
   }
 
   const resourceAttrList = getStringFromEnv('OTEL_RESOURCE_ATTRIBUTES');
-  if (resourceAttrList) {
+  const list = getStringListFromEnv('OTEL_RESOURCE_ATTRIBUTES');
+  if (list && list.length > 0) {
     config.resource.attributes_list = resourceAttrList;
+    config.resource.attributes = [];
+    for (let i = 0; i < list.length; i++) {
+      const element = list[i].split('=');
+      config.resource.attributes.push({
+        name: element[0],
+        value: element[1],
+        type: 'string',
+      });
+    }
   }
 
   const serviceName = getStringFromEnv('OTEL_SERVICE_NAME');